refactor(list-access): move utility methods to separate file
This way, they can be accessed by the write and read operations
