+++ src/dataflow/internal/process/functions/call/built-in/built-in-access.ts
@@ -11,18 +11,15 @@ import type { NodeId } from '../../../../../../r-bridge/lang-4.x/ast/model/proce
 import { dataflowLogger } from '../../../../../logger';
 import { RType } from '../../../../../../r-bridge/lang-4.x/ast/model/type';
 import { EdgeType } from '../../../../../graph/edge';
-import type { REnvironmentInformation } from '../../../../../environments/environment';
 import { makeAllMaybe, makeReferenceMaybe } from '../../../../../environments/environment';
 import type { ForceArguments } from '../common';
 import { BuiltIn } from '../../../../../environments/built-in';
 import { markAsAssignment } from './built-in-assignment';
 import { ReferenceType } from '../../../../../environments/identifier';
-import type { InGraphIdentifierDefinition } from '../../../../../environments/identifier';
-import { resolveByName } from '../../../../../environments/resolve-by-name';
 import type { ContainerIndicesCollection, ContainerParentIndex } from '../../../../../graph/vertex';
 import type { RArgument } from '../../../../../../r-bridge/lang-4.x/ast/model/nodes/r-argument';
 import { RoleInParent } from '../../../../../../r-bridge/lang-4.x/ast/model/processing/role';
-import type { Base } from '../../../../../../r-bridge/lang-4.x/ast/model/model';
+import { filterIndices, resolveSingleIndex } from '../../../../../../util/list-access';
 
 interface TableAssignmentProcessorMarker {
 	definitionRootNodes: NodeId[]
@@ -235,53 +232,6 @@ function processStringBasedAccess<OtherInfo>(
 	}
 
 	return fnCall;
-}
-
-/**
- * Resolves {@link accessedArg} in the {@link environment} and filters its indices according to {@link accessArg}.
- * 
- * @param accessedArg - The argument to resolve
- * @param accessArg - The argument which is used to filter the indices
- * @param environment - The environment in which {@link accessedArg} is resolved
- * @returns The filtered {@link ContainerIndicesCollection} of the resolved {@link accessedArg}
- */
-function resolveSingleIndex<OtherInfo>(
-	accessedArg: RArgument<OtherInfo & ParentInformation>,
-	accessArg: Base<OtherInfo & ParentInformation>,
-	environment: REnvironmentInformation,
-): ContainerIndicesCollection {
-	const definitions = resolveByName(accessedArg.lexeme, environment);
-	const indicesCollection = definitions?.flatMap(def => (def as InGraphIdentifierDefinition)?.indicesCollection ?? []);
-	const accessedIndicesCollection = filterIndices(indicesCollection, accessArg);
-	return accessedIndicesCollection;
-}
-
-/**
- * Filters the single indices of the {@link indicesCollection} according to the lexeme of the {@link accessArg}.
- * 
- * @param indicesCollection - The {@link ContainerIndicesCollection} to filter
- * @param accessArg - The argument which is used to filter {@link indicesCollection}
- * @returns The filtered copy of {@link indicesCollection}
- */
-function filterIndices<OtherInfo>(
-	indicesCollection: ContainerIndicesCollection,
-	accessArg: Base<OtherInfo & ParentInformation>,
-): ContainerIndicesCollection {
-	let accessedIndicesCollection: ContainerIndicesCollection = undefined;
-	for(const indices of indicesCollection ?? []) {
-		const filteredIndices = indices.indices.filter(index => accessArg.lexeme === index.lexeme);
-
-		if(filteredIndices.length == 0) {
-			continue;
-		}
-
-		accessedIndicesCollection ??= [];
-		accessedIndicesCollection.push({
-			indices:     filteredIndices,
-			isContainer: indices.isContainer
-		});
-	}
-	return accessedIndicesCollection;
 }
 
 /**
